requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pandas
plotly-express
python-decouple   # for env vars
PyJWT             # for decoding your Ionic JWTs
